+++ src/Model/Bcs/CreateShipment/RequestType/Economy.php
@@ -1,0 +1,24 @@
+<?php
+
+/**
+ * See LICENSE.md for license details.
+ */
+
+declare(strict_types=1);
+
+namespace Dhl\Sdk\Paket\Bcs\Model\Bcs\CreateShipment\RequestType;
+
+class Economy
+{
+    /**
+     * Indicates, if the option is on/off.
+     *
+     * @var int $active "0" or "1"
+     */
+    protected $active;
+
+    public function __construct(bool $active)
+    {
+        $this->active = (int) $active;
+    }
+}