+++ test/Provider/RequestData/AbstractRequestData.php
@@ -39,7 +39,6 @@ abstract class AbstractRequestData
      * @see ShipmentOrderRequestBuilderInterface::REQUEST_TYPE_REST
      *
      * @param ShipmentOrderRequestBuilderInterface $builder
-     * @param string $requestType
      * @param array $replace
      * @return ShipmentOrderType|\JsonSerializable
      * @throws RequestValidatorException